feature to source .tmux-restore in restored directory
Adding the ability to source a file on restore. If a pane is being
restored for a given directory and that directory contains a file named
.tmux-restore it will be sourced before restoring any running commands.
This is useful when working with projects that require environment setup
(ex: Andorid Open Source requires . build/envsetup.sh, projects with
python virtualenv, etc).